Fix generate shopping list
