+++ app/views/recipes/show.html.erb
@@ -61,9 +61,7 @@
 
   <div class="generate-shopping-list-container" id="pop-up"
     style="background-color:rgba(150,150,150, 0.7);
-    display: none;
-    visibility: hidden;
-    top: -10025%;
+    top: -100%;
     left:25%;
     width:50vw;
     height:50vh;
